+++ src/components/concepts/CodeView.svelte
@@ -46,7 +46,8 @@
             on:keydown={(event) =>
                 event.key === 'c' && (event.ctrlKey || event.metaKey)
                     ? copy()
-                    : undefined}><RootView {node} {inline} {spaces} /></div
+                    : undefined}
+            ><RootView {node} {inline} {spaces} preferred={true} /></div
         >{#if type && concept}&nbsp;<TypeView
                 {type}
                 context={concept.context}